+++ RELEASES.md
@@ -8,6 +8,7 @@ Version 0.9.1 (2022-12-XX)
 * Hyundai Tucson 2022-23 support
 * Kia Sorento 2022-23 support thanks to sunnyhaibin!
 * Kia Sorento Plug-in Hybrid 2022 support thanks to sunnyhaibin!
+* Volkswagen Crafter and MAN TGE 2017-23 support thanks to jyoung8607!
 
 Version 0.9.0 (2022-11-21)
 ========================